# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: CPEC-U Academic Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/assignments:
    get:
      operationId: listAssignments
      tags: [admin]
      summary: List all teacher-subject-class assignments
      responses:
        "200":
          description: Assignments list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeacherAssignment"
    post:
      operationId: createAssignment
      tags: [admin]
      summary: Assign a teacher to a subject in a class
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssignmentRequest"
      responses:
        "201":
          description: Assignment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeacherAssignment"
  /admin/assignments/{id}:
    delete:
      operationId: deleteAssignment
      tags: [admin]
      summary: Remove a teacher assignment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Assignment removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── Planificateur: Rooms ────────────────────────────────────────────
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    TeacherAssignment:
      type: object
      properties:
        id:
          type: integer
        teacherId:
          type: integer
        teacherName:
          type: string
        subjectId:
          type: integer
        subjectName:
          type: string
        coefficient:
          type: number
        classId:
          type: integer
        className:
          type: string
        semesterId:
          type: integer
        semesterName:
          type: string
      required: [id, teacherId, teacherName, subjectId, subjectName, coefficient, classId, className, semesterId, semesterName]
    CreateAssignmentRequest:
      type: object
      properties:
        teacherId:
          type: integer
        subjectId:
          type: integer
        classId:
          type: integer
        semesterId:
          type: integer
      required: [teacherId, subjectId, classId, semesterId]
...